Add seg control, refractor
