+++ berkeley-mobile/Resources/Covid/CovidResourceViewController.swift
@@ -1,0 +1,18 @@
+//
+//  CovidResourceViewController.swift
+//  berkeley-mobile
+//
+//  Created by Oscar Bjorkman on 10/31/20.
+//  Copyright © 2020 ASUC OCTO. All rights reserved.
+//
+
+import UIKit
+
+class CovidResourceViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+
+}
